+++ packages/system/src/index.ts
@@ -1,6 +1,7 @@
 import type {} from 'mongodb'
 import type {} from '@sonata-api/validation'
 export * from './collections'
+export * from './utils'
 
 export * as collections from './collections'
 export * as functions from './functions'